change ports: 5054 -> 15421, 8081 -> 15422
Both 5054 and 8081 are in a "busy zone" for ports, changing them to the
less occupied ports 15421/15422 makes it less likely another application
will be already using those ports (as already happening on the steam
deck).

Available ports were found with a combination of looking up
https://en.wikipedia.org/wiki/List_of_TCP_and_UDP_port_numbers and
searching "port NUMBER" and looking up the speedguide.net information on
the port.